fixed signed/unsigned warning
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2967 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/core/CLucene/search/PrefixQuery.cpp
@@ -94,7 +94,7 @@ CL_NS_DEF(search)
       const TCHAR* prefixField = prefix->field();
       const TCHAR* tmp;
       size_t i;
-	  int32_t prefixLen = prefix->textLength();
+	size_t prefixLen = prefix->textLength();
       do {
         lastTerm = enumerator->term();
 		if (lastTerm != NULL &&
@@ -235,7 +235,7 @@ BitSet* PrefixFilter::bits( IndexReader* reader )
     const TCHAR* prefixField = prefix->field();
     const TCHAR* tmp;
     size_t i;
-    int32_t prefixLen = prefix->textLength();
+    size_t prefixLen = prefix->textLength();
 	Term* lastTerm = NULL;
     
 	try{
